services/word-to-txt/requirements.txt
aiofiles==24.1.0
aio-pika==9.5.4
minio==7.2.15
python-docx==1.1.2

